requirements.txt
Flask==3.1.0
google-cloud-texttospeech==2.26.0
google-generativeai==0.3.2
python-dotenv==1.0.0

